+++ e2e/testconfig/testconfig.go
@@ -1,10 +1,21 @@
 package testconfig
 
 import (
+	"encoding/json"
 	"fmt"
 	"os"
 
+	"github.com/cosmos/cosmos-sdk/codec"
+	simappparams "github.com/cosmos/cosmos-sdk/simapp/params"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
+	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+	govv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
 	"github.com/strangelove-ventures/ibctest/ibc"
+	tmjson "github.com/tendermint/tendermint/libs/json"
+	tmtypes "github.com/tendermint/tendermint/types"
+
+	"github.com/cosmos/ibc-go/e2e/testvalues"
 )
 
 const (
@@ -90,6 +101,22 @@ func FromEnv() TestConfig {
 	}
 }
 
+func GetChainATag() string {
+	chainATag, ok := os.LookupEnv(ChainATagEnv)
+	if !ok {
+		panic(fmt.Sprintf("no environment variable specified for %s", ChainATagEnv))
+	}
+	return chainATag
+}
+
+func GetChainBTag() string {
+	chainBTag, ok := os.LookupEnv(ChainBTagEnv)
+	if !ok {
+		return GetChainATag()
+	}
+	return chainBTag
+}
+
 // ChainOptions stores chain configurations for the chains that will be
 // created for the tests. They can be modified by passing ChainOptionConfiguration
 // to E2ETestSuite.GetChains.
@@ -132,5 +159,54 @@ func newDefaultSimappConfig(cc ChainConfig, name, chainID, denom string) ibc.Cha
 		GasAdjustment:  1.3,
 		TrustingPeriod: "508h",
 		NoHostMount:    false,
+		ModifyGenesis:  defaultModifyGenesis(denom),
+	}
+}
+
+// defaultModifyGenesis will only modify governance params to ensure the voting period and minimum deposit
+// are functional for e2e testing purposes.
+func defaultModifyGenesis(denom string) func([]byte) ([]byte, error) {
+	return func(genbz []byte) ([]byte, error) {
+		genDoc, err := tmtypes.GenesisDocFromJSON(genbz)
+		if err != nil {
+			return nil, fmt.Errorf("failed to unmarshal genesis bytes into genesis doc: %w", err)
+		}
+
+		var appState genutiltypes.AppMap
+		if err := json.Unmarshal(genDoc.AppState, &appState); err != nil {
+			return nil, fmt.Errorf("failed to unmarshal genesis bytes into app state: %w", err)
+		}
+
+		cfg := simappparams.MakeTestEncodingConfig()
+		govv1beta1.RegisterInterfaces(cfg.InterfaceRegistry)
+		cdc := codec.NewProtoCodec(cfg.InterfaceRegistry)
+
+		govGenesisState := &govv1beta1.GenesisState{}
+		if err := cdc.UnmarshalJSON(appState[govtypes.ModuleName], govGenesisState); err != nil {
+			return nil, fmt.Errorf("failed to unmarshal genesis bytes into gov genesis state: %w", err)
+		}
+
+		// set correct minimum deposit using configured denom
+		govGenesisState.DepositParams.MinDeposit = sdk.NewCoins(sdk.NewCoin(denom, govv1beta1.DefaultMinDepositTokens))
+		govGenesisState.VotingParams.VotingPeriod = testvalues.VotingPeriod
+
+		govGenBz, err := cdc.MarshalJSON(govGenesisState)
+		if err != nil {
+			return nil, fmt.Errorf("failed to marshal gov genesis state: %w", err)
+		}
+
+		appState[govtypes.ModuleName] = govGenBz
+
+		genDoc.AppState, err = json.Marshal(appState)
+		if err != nil {
+			return nil, err
+		}
+
+		bz, err := tmjson.MarshalIndent(genDoc, "", "  ")
+		if err != nil {
+			return nil, err
+		}
+
+		return bz, nil
 	}
 }